+++ daemon/app/ghc-specter-daemon/Util.hs
@@ -1,30 +1,25 @@
 {-# LANGUAGE ForeignFunctionInterface #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module Util
-  ( demoLinePlots,
-    demoTables,
-    finalize,
+  ( finalize,
     initialize,
+    makeSparkline,
     showFramerate,
   )
 where
 
-import Control.Monad.Extra (whenM, whileM)
+import Control.Monad.Extra (whenM)
 import Data.Bits ((.|.))
-import Data.Foldable (for_, traverse_)
-import Data.IORef (IORef, modifyIORef', newIORef, readIORef)
 import Data.String (IsString (..))
 import FFICXX.Runtime.Cast (FPtr (..))
 import FFICXX.Runtime.TH (IsCPrimitive (..), TemplateParamInfo (..))
 import Foreign.C.String (CString, newCString, withCString)
-import Foreign.C.Types (CBool (..), CDouble (..), CFloat, CInt (..), CUInt (..))
-import Foreign.Marshal.Alloc (alloca)
-import Foreign.Marshal.Array (allocaArray)
+import Foreign.C.Types (CDouble (..), CFloat, CInt (..))
 import Foreign.Marshal.Utils (fromBool, toBool)
-import Foreign.Ptr (Ptr, castPtr, nullPtr)
-import Foreign.Storable (Storable (..))
+import Foreign.Ptr (Ptr, nullPtr)
 import ImGui
 import ImGui.Enum
 import ImGui.ImGuiIO.Implementation
@@ -32,15 +27,12 @@ import ImGui.ImGuiIO.Implementation
     imGuiIO_ConfigFlags_set,
     imGuiIO_Framerate_get,
   )
-import ImGui.ImVec2.Implementation (imVec2_x_get, imVec2_y_get)
-import ImGui.ImVec4.Implementation (imVec4_w_get, imVec4_x_get, imVec4_y_get, imVec4_z_get)
 import ImPlot qualified
 import ImPlot.Enum
 import ImPlot.TH qualified as TH
 import ImPlot.Template
 import STD.Deletable (delete)
 import System.IO.Unsafe (unsafePerformIO)
-import System.Random (randomRIO)
 import Text.Printf (printf)
 
 instance IsString CString where
@@ -84,33 +76,10 @@ TH.genPlotLine1InstanceFor
 
 showFramerate :: ImGuiIO -> IO ()
 showFramerate io = do
-  begin ("Framerate monitor" :: CString) nullPtr
+  _ <- begin ("Framerate monitor" :: CString) nullPtr
   framerate :: Float <- realToFrac <$> imGuiIO_Framerate_get io
   withCString (printf "Application average %.3f ms/frame (%.1f FPS)" (1000.0 / framerate) framerate) $ \c_str ->
     textUnformatted c_str
-  end
-
-demoLinePlots :: (Ptr CFloat, Ptr CFloat) -> (Ptr CDouble, Ptr CDouble) -> IO ()
-demoLinePlots (px1, py1) (px2, py2) = do
-  begin ("Line plots" :: CString) nullPtr
-  whenM (toBool <$> ImPlot.beginPlot_ ("Line Plots" :: CString)) $ do
-    ImPlot.setupAxes
-      ("x" :: CString)
-      ("y" :: CString)
-      (fromIntegral (fromEnum ImPlotAxisFlags_None))
-      (fromIntegral (fromEnum ImPlotAxisFlags_None))
-    t <- getTime
-    for_ [0 .. 1000] $ \i -> do
-      let x = fromIntegral i * 0.001
-      pokeElemOff px1 i x
-      pokeElemOff py1 i (0.5 + 0.5 * sin (50.0 * (x + realToFrac t / 10.0)))
-    plotLine "f(x)" px1 py1 1001
-    for_ [0 .. 19] $ \i -> do
-      let x = fromIntegral i / 19.0
-      pokeElemOff px2 i x
-      pokeElemOff py2 i (x * x)
-    plotLine "g(x)" px2 py2 20
-    ImPlot.endPlot
   end
 
 makeSparkline :: Ptr CFloat -> Int -> IO ()
@@ -129,37 +98,6 @@ makeSparkline pdat offset = do
     ImPlot.endPlot
   delete spark_size
   delete zero
-
-demoTables :: IORef Int -> Ptr CFloat -> IO ()
-demoTables ref_offset pdat = do
-  begin ("Table of plots" :: CString) nullPtr
-  let flags =
-        fromIntegral $
-          fromEnum ImGuiTableFlags_BordersOuter
-            .|. fromEnum ImGuiTableFlags_BordersV
-            .|. fromEnum ImGuiTableFlags_RowBg
-            .|. fromEnum ImGuiTableFlags_Resizable
-            .|. fromEnum ImGuiTableFlags_Reorderable
-  whenM (toBool <$> beginTable ("##table" :: CString) 3 (fromIntegral flags)) $ do
-    tableSetupColumn ("Electrode" :: CString) (fromIntegral (fromEnum ImGuiTableColumnFlags_WidthFixed)) 75.0
-    tableSetupColumn ("Voltage" :: CString) (fromIntegral (fromEnum ImGuiTableColumnFlags_WidthFixed)) 75.0
-    tableSetupColumn_ ("EMG Signal" :: CString)
-    tableHeadersRow
-    offset <- readIORef ref_offset
-    for_ [0 .. 9] $ \(row :: Int) -> do
-      let offset' = (offset + row * 10) `mod` 100
-      tableNextRow 0
-      tableSetColumnIndex 0
-      textUnformatted (fromString (printf "EMG %d" row) :: CString)
-      tableSetColumnIndex 1
-      val :: Float <- realToFrac <$> peekElemOff pdat offset'
-      textUnformatted (fromString (printf "%.3f V" val) :: CString)
-      tableSetColumnIndex 2
-      pushID (fromIntegral row)
-      makeSparkline pdat offset'
-      popID
-    endTable
-  end
 
 initialize :: IO (ImGuiContext, ImGuiIO, GLFWwindow)
 initialize = do
@@ -190,8 +128,8 @@ initialize = do
   styleColorsLight
 
   -- Setup Platform/Renderer backends
-  imGui_ImplGlfw_InitForOpenGL window (fromBool True)
-  imGui_ImplOpenGL3_Init glsl_version
+  _ <- imGui_ImplGlfw_InitForOpenGL window (fromBool True)
+  _ <- imGui_ImplOpenGL3_Init glsl_version
 
   -- Enable Keyboard Controls and Gamepad Controls
   io <- getIO
